+++ python/cugraph/__init__.py
@@ -43,7 +43,7 @@ from cugraph.utilities import device_of_gpu_pointer
 
 from cugraph.snmg.link_analysis.mg_pagerank import mg_pagerank
 
-from cugraph.pregel.traversal import bfs_df_pregel
+from cugraph.bsp.traversal import bfs_df_pregel
 
 from cugraph.proto.components import strong_connected_component
 from cugraph.proto.structure import find_bicliques
