+++ python/cugraph/bsp/traversal/__init__.py
@@ -11,5 +11,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from cugraph.pregel.traversal.bfs_bsp import bfs_df_pregel
+from cugraph.bsp.traversal.bfs_bsp import bfs_df_pregel
